requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
sentence-transformers>=2.3.0
pdfplumber>=0.10.0
python-docx>=1.1.0
nltk>=3.8.0
